Added required casts to fix 64-bit compilation errors
Reviewed-on: https://chromium-review.googlesource.com/244510
Tested-by: Cooper Partin <coopp@microsoft.com>

+++ src/libANGLE/Context.cpp
@@ -230,7 +230,7 @@ GLsync Context::createFenceSync()
 {
     GLuint handle = mResourceManager->createFenceSync();
 
-    return reinterpret_cast<GLsync>(handle);
+    return reinterpret_cast<GLsync>(static_cast<uintptr_t>(handle));
 }
 
 GLuint Context::createVertexArray()
